research_analyzer_last/requirements.txt
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
fake-useragent==1.4.0
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
SQLAlchemy==2.0.23
aiosmtplib==2.0.2
fpdf2==2.7.5
matplotlib==3.7.1
scholarly==1.7.11
torch==2.1.0
transformers==4.35.0
sentencepiece==0.1.99
rich==13.7.0
flask==3.0.0
qrcode==7.4.2
schedule==1.2.0